Updated seam carving

+++ CS155Project4-WangTiles/tiles.h
@@ -63,7 +63,7 @@ class Tiles
         void initTextures(Image* src);     // sample patches from picture, store in himage_, vimage_
     
         int getRandomTile(int up, int left);  // randomly select a tile with matching NW edge
-        int colorDiff(double r, double g, double b);  // color difference
+        double colorDiff(double r, double g, double b);  // color difference
         vector<int> seamCarving(Image* src1, Image* srr2);  // color difference
         Image* genCollage(int n, int e, int s, int w);
